Moved strings and activity declaration
