+++ androidHackathonApp/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-<string name="app_name">CommunityKitchen</string>
+    <string name="app_name">CommunityKitchen</string>
     <string name="subheading">Add a meal to help out someone in your local community or browse the meals that others have added</string>
     <string name="add_meal_title">Add a Meal</string>
     <string name="find_meal_title">Find a Meal</string>
