+++ lib-find-meal/src/main/AndroidManifest.xml
@@ -1,2 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.kinandcarta.lib.find.meal" />
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.kinandcarta.lib.find.meal">
+
+    <application>
+        <activity
+            android:name=".view.FindMealActivity"
+            android:label="@string/find_meal_activity_header" />
+    </application>
+
+</manifest>
