+++ lib-find-meal/src/main/res/layout/find_meal_activity.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.kinandcarta.lib.find.meal.FindMealActivity">
+    tools:context="com.kinandcarta.lib.find.meal.view.FindMealActivity">
 
 <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
     android:id="@+id/swipeContainer"
