+++ lib-find-meal/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="find_meal_activity_header">This is the find meal activity screen</string>
+</resources>
